+++ POS.Backend/API_Common/Const/GiftStatusConst.cs
@@ -1,8 +1,0 @@
-﻿
-namespace TCX.API.Common.Constants
-{
-    public class GiftStatusConst
-    {
-        public static readonly string[] CheckStatusCreateGift = { "CHECK", "CREATE" };
-    }
-}