+++ POS.Backend/API_Common/Models/Odoo/ResponseUpdStt.cs
@@ -1,6 +1,0 @@
-﻿namespace VCM.POSBLUE.API.Controllers
-{
-    internal class ResponseUpdStt
-    {
-    }
-}